feat: move company api to /admin, authentication
