+++ backend/src/admin/hooks/employees.tsx
@@ -28,7 +28,8 @@ export const useEmployees = (
     const fetchEmployees = async () => {
       try {
         const response = await fetch(
-          `/companies/${companyId}/employees` + (query ? `?${filterQuery}` : "")
+          `/admin/companies/${companyId}/employees` +
+            (query ? `?${filterQuery}` : "")
         );
         const result = await response.json();
         setData(result);
@@ -62,7 +63,7 @@ export const useCreateEmployee = (
     setError(null);
 
     try {
-      const response = await fetch(`/companies/${companyId}/employees`, {
+      const response = await fetch(`/admin/companies/${companyId}/employees`, {
         method: "POST",
         headers: {
           "Content-Type": "application/json",
@@ -106,7 +107,7 @@ export const useUpdateEmployee = (
 
     try {
       const response = await fetch(
-        `/companies/${companyId}/employees/${employeeId}`,
+        `/admin/companies/${companyId}/employees/${employeeId}`,
         {
           method: "POST",
           headers: {
@@ -152,7 +153,7 @@ export const useDeleteEmployee = (
 
     try {
       const response = await fetch(
-        `/companies/${companyId}/employees/${employeeId}`,
+        `/admin/companies/${companyId}/employees/${employeeId}`,
         {
           method: "DELETE",
         }
