+++ backend/src/api/admin/companies/[id]/employees/route.ts
@@ -1,7 +1,7 @@
 import { createCustomersWorkflow } from "@medusajs/core-flows";
 import { ContainerRegistrationKeys } from "@medusajs/framework/utils";
 import { MedusaRequest, MedusaResponse } from "@medusajs/medusa";
-import { createEmployeesWorkflow } from "../../../../workflows/employee/workflows";
+import { createEmployeesWorkflow } from "../../../../../workflows/employee/workflows";
 import { CreateEmployeeType, GetEmployeeParamsType } from "./validators";
 
 export const GET = async (
