+++ backend/src/api/admin/companies/query-config.ts
