+++ backend/src/api/admin/companies/validators.ts
