+++ backend/src/api/admin/middlewares.ts
@@ -1,0 +1,8 @@
+import { MiddlewareRoute } from "@medusajs/medusa";
+import { employeesMiddlewares } from "./companies/[id]/employees/middlewares";
+import { companiesMiddlewares } from "./companies/middlewares";
+
+export const adminMiddlewares: MiddlewareRoute[] = [
+  ...companiesMiddlewares,
+  ...employeesMiddlewares,
+];
