+++ backend/src/api/middlewares.ts
@@ -1,7 +1,7 @@
 import { defineMiddlewares } from "@medusajs/medusa";
 import { adminQuotesMiddlewares } from "./admin/quotes/middlewares";
-import { employeesMiddlewares } from "./companies/[id]/employees/middlewares";
-import { companiesMiddlewares } from "./companies/middlewares";
+import { employeesMiddlewares } from "./admin/companies/[id]/employees/middlewares";
+import { companiesMiddlewares } from "./admin/companies/middlewares";
 import { quotesMiddlewares } from "./customers/quotes/middlewares";
 
 export default defineMiddlewares({
